+++ convey/identifier.py
@@ -271,7 +271,7 @@ class Identifier:
             return Types.plaintext
         return fitting_type
 
-    def get_fitting_source_i(self, target_type: Type, try_hard=False):
+    def get_fitting_source_i(self, target_type: Type, try_hard=False) -> List[int]:
         """ Get list of source_i that may be of such a field type that new_field would be computed effectively.
             Note there is no fitting column for TypeGroup.custom, if you try_hard, you receive first column as a plaintext.
 
@@ -425,7 +425,7 @@ class Identifier:
         """
         Useful for parsing user input COLUMN from the CLI args.
         :type column: object Either column ID (ex "1" points to column index 0), position from the right (-1, ...),
-                        or an exact column name or the field
+                        desired type or an exact column name or the field
         :type check: If not False and not found, error is raised and quit. If str, this string will be included in the error text.
         :rtype: int Either column_i or None if not found.
         """
@@ -437,19 +437,13 @@ class Identifier:
             source_col_i = i - 1 if i > 0 else i  # ID=1 -> col_i=0 (first column), ID=-1 -> col_i=-1 (last added)
         elif column in self.parser.first_line_fields:  # exact column name
             source_col_i = self.parser.first_line_fields.index(column)
-        else:
+        else:  # as of Python3.8, shorten to elif searched_type :=...
             searched_type = Types.find_type(column)  # get field by its type
             if searched_type:
-                reserve = None
-                for f in self.parser.fields:
-                    if f.type == searched_type:
-                        source_col_i = f.col_i
-                    elif searched_type in f.possible_types and not reserve:
-                        reserve = f.col_i
-                if not source_col_i:
-                    source_col_i = reserve
+                source_col_i = (next((f.col_i for f in self.parser.fields if f.type == searched_type), None) or
+                                next((f.col_i for f in self.parser.fields if searched_type in f.possible_types), None))
         if check and (source_col_i is None or len(self.parser.fields) <= source_col_i):
             logger.error(f"Cannot identify COLUMN {column}" + (" " + check if type(check) is str else "") +
-                         ", put there an exact column name or the numerical order starting with 1.")
+                         ", put there an exact column name, its type, the numerical order starting with 1, or with -1.")
             quit()
         return source_col_i